Add non gui functionality to the STM32F Programmer
In preparation for adding the STM32F example script that does not
use the api object but the cw4.0 api for the scope instead, changes
had to be made to the STM32FProgrammer and the stm32 usb interface.
The implementation is meant to stay consistant as possible with the
XMEGA Programmer.