Fix for aux module preview not updating
The problem ocurs when removing a aux module and rerunnning
the script for the addition of an updated aux module the underlieing
logic would change but the code preview would not. This was cofusing.

It turns out the underlieing source inspection module uses a cache.
So the fix is currently clearing that cache. There may be a better way
but this way is short and sweet and does not have negative effects
on performance since we read no significant length files where caches
would be beneficial.