SimpleSerial: Add support for 'noflush' kwarg
Although real-world targets won't be speaking the SimpleSerial protocol
used for the various tutorials and exercises, this class can still be
leveraged to communicate with a target's UART via read() and write().

However, _con() currently writes to the UART (before a caller can change
the baud rate!), which may inadvertently interfere with our
beloved ChipWhisper-wielding mischiefmaker may be trying to accomplish.

The 'noflush' kwarg can now suppress this behavior.

Example:

	target = cw.targets(scope, noflush=True)